+++ tests/helpers/healthevent.go
@@ -15,6 +15,8 @@
 package helpers
 
 import (
+	"context"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
@@ -22,8 +24,173 @@ import (
 	"os"
 	"strings"
 	"sync"
+	"testing"
 	"time"
+
+	"github.com/stretchr/testify/require"
 )
+
+type HealthEventTemplate struct {
+	Version             int                  `json:"version"`
+	Agent               string               `json:"agent"`
+	ComponentClass      string               `json:"componentClass,omitempty"`
+	CheckName           string               `json:"checkName"`
+	IsFatal             bool                 `json:"isFatal"`
+	IsHealthy           bool                 `json:"isHealthy"`
+	Message             string               `json:"message"`
+	RecommendedAction   int                  `json:"recommendedAction,omitempty"`
+	ErrorCode           []string             `json:"errorCode,omitempty"`
+	EntitiesImpacted    []EntityImpacted     `json:"entitiesImpacted,omitempty"`
+	Metadata            map[string]string    `json:"metadata,omitempty"`
+	QuarantineOverrides *QuarantineOverrides `json:"quarantineOverrides,omitempty"`
+	NodeName            string               `json:"nodeName"`
+}
+
+type EntityImpacted struct {
+	EntityType  string `json:"entityType"`
+	EntityValue string `json:"entityValue"`
+}
+
+type QuarantineOverrides struct {
+	Force bool `json:"force"`
+}
+
+func NewHealthEvent(nodeName string) *HealthEventTemplate {
+	return &HealthEventTemplate{
+		Version:        1,
+		Agent:          "gpu-health-monitor",
+		ComponentClass: "GPU",
+		CheckName:      "GpuXidError",
+		IsFatal:        true,
+		IsHealthy:      false,
+		NodeName:       nodeName,
+		EntitiesImpacted: []EntityImpacted{
+			{
+				EntityType:  "GPU",
+				EntityValue: "0",
+			},
+		},
+	}
+}
+
+func (h *HealthEventTemplate) WithAgent(agent string) *HealthEventTemplate {
+	h.Agent = agent
+	return h
+}
+
+func (h *HealthEventTemplate) WithCheckName(checkName string) *HealthEventTemplate {
+	h.CheckName = checkName
+	return h
+}
+
+func (h *HealthEventTemplate) WithErrorCode(codes ...string) *HealthEventTemplate {
+	h.ErrorCode = codes
+	return h
+}
+
+func (h *HealthEventTemplate) WithComponentClass(class string) *HealthEventTemplate {
+	h.ComponentClass = class
+	return h
+}
+
+func (h *HealthEventTemplate) WithEntity(entityType, entityValue string) *HealthEventTemplate {
+	h.EntitiesImpacted = append(h.EntitiesImpacted, EntityImpacted{
+		EntityType:  entityType,
+		EntityValue: entityValue,
+	})
+	return h
+}
+
+func (h *HealthEventTemplate) WithFatal(isFatal bool) *HealthEventTemplate {
+	h.IsFatal = isFatal
+	return h
+}
+
+func (h *HealthEventTemplate) WithHealthy(isHealthy bool) *HealthEventTemplate {
+	h.IsHealthy = isHealthy
+	return h
+}
+
+func (h *HealthEventTemplate) WithMessage(message string) *HealthEventTemplate {
+	h.Message = message
+	return h
+}
+
+func (h *HealthEventTemplate) WithForceOverride() *HealthEventTemplate {
+	h.QuarantineOverrides = &QuarantineOverrides{Force: true}
+	if h.Metadata == nil {
+		h.Metadata = make(map[string]string)
+	}
+	h.Metadata["creator_id"] = "test"
+	return h
+}
+
+func (h *HealthEventTemplate) WithMetadata(key, value string) *HealthEventTemplate {
+	if h.Metadata == nil {
+		h.Metadata = make(map[string]string)
+	}
+	h.Metadata[key] = value
+	return h
+}
+
+func (h *HealthEventTemplate) WriteToTempFile() (string, error) {
+	tempFile, err := os.CreateTemp("", "health-event-*.json")
+	if err != nil {
+		return "", fmt.Errorf("failed to create temp file: %w", err)
+	}
+
+	content, err := json.MarshalIndent(h, "", "    ")
+	if err != nil {
+		tempFile.Close()
+		os.Remove(tempFile.Name())
+		return "", fmt.Errorf("failed to marshal health event: %w", err)
+	}
+
+	if _, err := tempFile.Write(content); err != nil {
+		tempFile.Close()
+		os.Remove(tempFile.Name())
+		return "", fmt.Errorf("failed to write to temp file: %w", err)
+	}
+
+	tempFile.Close()
+	return tempFile.Name(), nil
+}
+
+// SendHealthEventWithTemplate sends a health event and returns the temp file path.
+// The caller is responsible for cleaning up the temp file with defer os.Remove(tempFile).
+//
+// For automatic cleanup, use SendHealthEventWithAutoCleanup instead.
+func SendHealthEventWithTemplate(nodeName string, event *HealthEventTemplate) (string, error) {
+	tempFile, err := event.WriteToTempFile()
+	if err != nil {
+		return "", err
+	}
+
+	err = SendHealthEventsToNodes([]string{nodeName}, tempFile)
+	if err != nil {
+		os.Remove(tempFile)
+		return "", err
+	}
+
+	return tempFile, nil
+}
+
+// SendHealthEventWithAutoCleanup sends a health event and automatically cleans up the temp file.
+// Use this when you don't need to control the temp file lifetime.
+func SendHealthEventWithAutoCleanup(nodeName string, event *HealthEventTemplate) error {
+	tempFile, err := event.WriteToTempFile()
+	if err != nil {
+		return err
+	}
+	defer os.Remove(tempFile)
+
+	err = SendHealthEventsToNodes([]string{nodeName}, tempFile)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
 
 // SendHealthEventsToNodes sends health events from the specified `eventFilePath` to all nodes listed in `nodeNames` concurrently.
 func SendHealthEventsToNodes(nodeNames []string, eventFilePath string) error {
@@ -68,4 +235,31 @@ func SendHealthEventsToNodes(nodeNames []string, eventFilePath string) error {
 	wg.Wait()
 
 	return errors.Join(errs...)
+}
+
+// SendHealthEvent is a test helper that sends a health event with logging and automatic error handling.
+// Returns the temp file path that the caller should clean up with defer os.Remove(tempFile).
+func SendHealthEvent(ctx context.Context, t *testing.T, event *HealthEventTemplate) string {
+	t.Logf("Sending health event to node %s: checkName=%s, isFatal=%v",
+		event.NodeName, event.CheckName, event.IsFatal)
+	tempFile, err := SendHealthEventWithTemplate(event.NodeName, event)
+	require.NoError(t, err)
+	t.Logf("Health event sent successfully")
+	return tempFile
+}
+
+// SendHealthyEvent sends a generic healthy event to clear any health issues on a node.
+func SendHealthyEvent(ctx context.Context, t *testing.T, nodeName string) {
+	t.Logf("Sending generic healthy event to node %s", nodeName)
+	event := NewHealthEvent(nodeName).
+		WithHealthy(true).
+		WithFatal(false).
+		WithMessage("No health failures").
+		WithComponentClass("GPU").
+		WithErrorCode("")
+
+	event.ErrorCode = nil
+
+	tempFile := SendHealthEvent(ctx, t, event)
+	defer os.Remove(tempFile)
 }
